requirements.txt
python-telegram-bot==22.5
python-dotenv==1.0.0
sqlite3
pandas==2.1.4
numpy==1.24.3
cryptography==41.0.7
langdetect==1.0.9
googletrans==4.0.0rc1
transformers==4.35.2
torch==2.1.1
requests==2.31.0
schedule==1.2.0
python-dateutil==2.8.2
